requirements.txt
setuptools
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.4
django-guardian==2.4.0
graphene-django